# docs/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "O-Week v1.1 API",
    "version": "1.1.0",
    "description": "Formal User-based API. Authentication uses the HttpOnly owk_session cookie."
  },
  "servers": [{ "url": "/api/v1" }],
...
  "paths": {
...
    "/submissions/{section}/submit": { "post": { "tags": ["Student"], "operationId": "submitSection", "security": [{ "sessionCookie": [] }], "parameters": [{ "$ref": "#/components/parameters/Section" }, { "$ref": "#/components/parameters/IdempotencyKey" }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Submit" } } } }, "responses": { "200": { "$ref": "#/components/responses/Success" }, "403": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "422": { "$ref": "#/components/responses/Error" } } } },
...
    "/admin/settings/apply-preset": { "post": { "tags": ["Admin"], "operationId": "applyAdminPreset", "security": [{ "sessionCookie": [] }], "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }], "responses": { "200": { "$ref": "#/components/responses/Success" }, "409": { "$ref": "#/components/responses/Error" } } } },
...
    "/admin/groups": { "post": { "tags": ["Admin"], "operationId": "createAdminGroup", "security": [{ "sessionCookie": [] }], "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }], "responses": { "200": { "$ref": "#/components/responses/Success" }, "409": { "$ref": "#/components/responses/Error" } } } },
    "/admin/groups/{id}": { "patch": { "tags": ["Admin"], "operationId": "renameAdminGroup", "security": [{ "sessionCookie": [] }], "parameters": [{ "$ref": "#/components/parameters/GroupId" }, { "$ref": "#/components/parameters/IdempotencyKey" }], "responses": { "200": { "$ref": "#/components/responses/Success" }, "404": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" } } }, "delete": { "tags": ["Admin"], "operationId": "deleteAdminGroup", "description": "Unassigns every member before deleting the group. Senior accounts become fail-closed until assigned to another group.", "security": [{ "sessionCookie": [] }], "parameters": [{ "$ref": "#/components/parameters/GroupId" }, { "$ref": "#/components/parameters/IdempotencyKey" }], "responses": { "200": { "$ref": "#/components/responses/Success" }, "404": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" } } } },
...
  },
  "components": {
    "securitySchemes": { "sessionCookie": { "type": "apiKey", "in": "cookie", "name": "owk_session" } },
    "parameters": {
...
      "Section": { "name": "section", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/Section" } },
...
      "GroupId": { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
      "IdempotencyKey": { "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "minLength": 1, "maxLength": 200 } }
    },
    "schemas": {
      "Section": { "type": "string", "enum": ["day1", "day3"] },
...
      "Submit": { "type": "object", "additionalProperties": false, "required": ["version", "confirm"], "properties": { "version": { "type": "integer", "minimum": 1 }, "confirm": { "const": true } } },
...
      "ErrorEnvelope": { "type": "object", "required": ["error", "requestId"], "properties": { "error": { "type": "object", "required": ["code", "message", "details"], "properties": { "code": { "type": "string" }, "message": { "type": "string" }, "details": { "type": "object" } } }, "requestId": { "type": "string", "pattern": "^req_" } } },
      "SuccessEnvelope": { "type": "object", "required": ["data", "requestId"], "properties": { "data": {}, "requestId": { "type": "string", "pattern": "^req_" } } }
    },
    "responses": {
      "Success": { "description": "Successful response", "headers": { "Cache-Control": { "schema": { "type": "string" } } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessEnvelope" } } } },
      "Error": { "description": "Stable error envelope", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } }
    }
  }
}